Refactor to use app module for FOLIO Migration Reports
