+++ plugins/folio/templates/folio/index.html
@@ -27,23 +27,23 @@
                 <li>
                     <strong>Instances</strong>
                     {% for row in info["sources"]["instances"] %}
-                    <a href="{{ url_for('FOLIO.source_record', iteration_id=dag_run_id, folio_type='instances', filename=row) }}" download>{{ row }}</a>
+                    <a href="{{ url_for('FOLIOMigrationReports.source_record', iteration_id=dag_run_id, folio_type='instances', filename=row) }}" download>{{ row }}</a>
                     {% endfor %}
                 </li>
                 <li>
                     <strong>Holdings</strong>
                     {% for row in info["sources"]["holdings"] %}
                     {% if "mhlds" in row %}
-                    <a href="{{ url_for('FOLIO.source_record', iteration_id=dag_run_id, folio_type='holdings', filename=row) }}" download>{{ row }}</a>
+                    <a href="{{ url_for('FOLIOMigrationReports.source_record', iteration_id=dag_run_id, folio_type='holdings', filename=row) }}" download>{{ row }}</a>
                     {% else %}
-                    <a href="{{ url_for('FOLIO.source_record', iteration_id=dag_run_id, folio_type='items', filename=row) }}" download>{{ row }}</a>
+                    <a href="{{ url_for('FOLIOMigrationReports.source_record', iteration_id=dag_run_id, folio_type='items', filename=row) }}" download>{{ row }}</a>
                     {% endif %}
                     {% endfor %}    
                 </li>
                 <li>
                     <strong>Items</strong>
                     {% for row in info["sources"]["items"] %}
-                    <a href="{{ url_for('FOLIO.source_record', iteration_id=dag_run_id, folio_type='items', filename=row) }}" download>{{ row }}</a>
+                    <a href="{{ url_for('FOLIOMigrationReports.source_record', iteration_id=dag_run_id, folio_type='items', filename=row) }}" download>{{ row }}</a>
                     {% endfor %} 
                 </li>
             </ul>
@@ -51,14 +51,14 @@
         <td>
             <ul>
             {% for report in info.get('reports') %}
-            <li><a href="{{ url_for('FOLIO.folio_report', iteration_id=dag_run_id, report_name=report) }}">{{ report }}</a></li>
+            <li><a href="{{ url_for('FOLIOMigrationReports.folio_report', iteration_id=dag_run_id, report_name=report) }}">{{ report }}</a></li>
             {% endfor %}
             </ul>
         </td>
         <td>
             <ul>
             {% for issues_log in info.get('data_issues') %}
-            <li><a href="{{ url_for('FOLIO.folio_data_issues', iteration_id=dag_run_id, log_name=issues_log) }}">{{ issues_log }}</a></li>
+            <li><a href="{{ url_for('FOLIOMigrationReports.folio_data_issues', iteration_id=dag_run_id, log_name=issues_log) }}">{{ issues_log }}</a></li>
             {% endfor %}
             </ul>
         </td>
@@ -67,25 +67,25 @@
                 <li>
                     <strong>Instances</strong>
                     {% for row in info["records"]["instances"] %}
-                        <a href="{{ url_for('FOLIO.folio_json_records', iteration_id=dag_run_id, filename=row) }}" download>{{ row }}</a>
+                        <a href="{{ url_for('FOLIOMigrationReports.folio_json_records', iteration_id=dag_run_id, filename=row) }}" download>{{ row }}</a>
                     {% endfor %}
                 </li>
                 <li>
                     <strong>Holdings</strong>
                     {% for row in info["records"]["holdings"] %}
-                    <a href="{{ url_for('FOLIO.folio_json_records', iteration_id=dag_run_id, filename=row) }}" download>{{ row }}</a>
+                    <a href="{{ url_for('FOLIOMigrationReports.folio_json_records', iteration_id=dag_run_id, filename=row) }}" download>{{ row }}</a>
                     {% endfor %}    
                 </li>
                 <li>
                     <strong>Items</strong>
                     {% for row in info["records"]["items"] %}
-                    <a href="{{ url_for('FOLIO.folio_json_records', iteration_id=dag_run_id, filename=row) }}" download>{{ row }}</a>
+                    <a href="{{ url_for('FOLIOMigrationReports.folio_json_records', iteration_id=dag_run_id, filename=row) }}" download>{{ row }}</a>
                     {% endfor %} 
                 </li>
                 <li>
                     <strong>Source Record Storage</strong>
                     {% for row in info["records"]["srs"] %}
-                    <a href="{{ url_for('FOLIO.folio_json_records', iteration_id=dag_run_id, filename=row) }}" download>{{ row }}</a>
+                    <a href="{{ url_for('FOLIOMigrationReports.folio_json_records', iteration_id=dag_run_id, filename=row) }}" download>{{ row }}</a>
                     {% endfor %}
                 </li>
             </ul>
